moved queries from server to client
