+++ client/src/utils/queries.js
@@ -1,0 +1,25 @@
+import { gql } from "@apollo/client";
+
+export const QUERY_USER = gql`
+  query user($username: String!) {
+    user(username: $username) {
+      _id
+      username
+      email
+      raffleTickets {
+        ticketId
+      }
+    }
+  }
+`;
+
+export const QUERY_ME = gql`
+  {
+    me {
+      _id
+      username
+      email
+      
+    }
+  }
+`;
